requirements.txt
fastapi[standard]>=0.95.0
uvicorn>=0.22.0
langchain>=0.0.300
openai>=0.27.0
python-dotenv>=1.0.0
pytest>=7.0.0
requests>=2.28.0
sqlmodel>=0.0.8
SQLAlchemy>=1.4.0
